CI: Replace current Haskell test suite with new Docker-based approach
This is necessary due to the introduction of the restricted user, making
it essentially impossible to run locally at this point (unless you have
a user locally called restricted).

However, in the future when there will be ACLs applied to the restricted
users in the docker images, running it locally in any capacity outside
of the Docker image, would likely be cumbersome and annoying.