Refactor: Move `TIMEOUT` constant to runner module
Was located in Haskell module, however it makes more sense to make it a
'global' thing, rather than a module specific thing.

If we are interested in introducing language specific or exercise
specific timeouts, then it would make more sense to have it be a
parameter of the request, rather than a local constant anyhow.