Python: Add integration test case for limiting permissions
This is only two test cases for the places that matter the most imo.
Preventing writes to the /mozart directory, will prevent influence on
test runner jobs in weird scenarios. Preventing writes to /tmp will
prevent files from being created in a common and easy to access writable
folder (although there are of course other folders they may be able to
target).