Feat: Include cause of runtime error in JSON output
This modifies the exception handling in Haskell and Python to include
the exception stringified (with newlines escaped), so that it can be
included in the `details` field of a given test case.

This means that the underlying cause of the runtime error can now be
presented to the end-user as part of the feedback.