Fix: Correctly remove working directory on client-side cancellation
In axum each request has an associated Future, and if the request is
cancelled the Future is also cancelled. This would allow a scenario
where a client-side cancelled request would prevent the clean-up of the
unique working directory for that request.

This change makes it so that a middleware layer dissociates the handler
function Future with the underlying request, thereby removing the
effect of client side cancellation on the request handler.

This means that the clean-up will correctly run and the associated
request will have its directory removed, however, it may also lead to
more work (because it will likely run requests to end instead of failing
early).